requirement.txt
altgraph==0.17.2
asgiref==3.5.2
bcrypt==4.0.1
distlib==0.3.5
Django==4.0.6
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
docopt==0.6.2
emoji==2.2.0
et-xmlfile==1.1.0
filelock==3.7.1
future==0.18.2
gunicorn==21.2.0
Nuitka==1.0.8
num2words==0.5.12
openpyxl==3.1.0
packaging==23.1
pefile==2022.5.30
Pillow==9.4.0
platformdirs==2.5.2
psycopg2==2.9.3
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pynput==1.7.6
python-decouple==3.8
pytz==2023.3
pywin32-ctypes==0.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.15.1
virtualenvwrapper-win==1.2.7
whitenoise==6.5.0
